Foreground and background means now use a blurred version of the image, and when they display the number they display it in a cleaner fashion, though it is still to long to see.
Moved the ImageEnergies to an energy package. Added the Balloon + gradient energy which expands from the center if the contour lies on a region brighter than the foreground value. Changed the imageprocessor.filter(1) to .filter(ImageProcessor.FIND_EDGES) for clarity.

SnakeFrame: changed the way that the selected energies are acquired. Added a new row of buttons. Added those buttons and their actions. Explained in model.

SnakeImages Added a sigma value for the get averaged value. Previous version just used 1.01. Changed the way pixels are accessed, possibly slower, but more consistent. Uses the getInterpolatedValue which should be the same as the energies.

SnakeIO: Added a PWD since the latest version of linux forget what directory you were in and default to a directory that is broken on linux. (I have submitted a patch to openjdk concerning this issue.)  Added a new method so that there is a getOpenFile with a title argument. Also fixed in the images so that you get 'select image file' instead of 'select snake'.

SnakeListener: New cases to handle new buttons. Explained in model. Some of these cases include the modifiers and actions are affected accordingly.

SnakeModel: 
-Moved thread execution to a fixed thread pool instead of starting a new thread. 
-Changed the default gamma to be 400 since 40 has always been too small.
-Added a check to see if deformIterations<0, if it is then the program will attempt to deform until the user interrupts or the number of point in the snake does not change appreciable after 10 iterations.
-Changed the image intensity selection to include 3 selections. 
-Added a private class for modifying snakes on the event loop. It handles some exceptions.
-Added a track all frames method. It can be modified with CTRL and SHIFT. Just click the button and it tracks all frames forward. Click the button holding down shift, and it tracks all frames backwards. Click the button holding down CTRL+SHIFT it tracks forward then goes to the starting frame and tracks backwards.
-Track backwards, just like the normal track, but it moves back a frame.
-deform all frames. Goes through all frames the snake exists in and deforms it. If you hold down SHIFT or CTRL it will start at the current frame and go forward.
-time smooth snakes was going to be a smooth snake in time method, but I didn't get around to it. Instead it measures the FOREGROUND and BACKGROUND intensity of the BalloonGradient.
-getSigma() - getter of sigma so other functions can use it.
TwoDContourDeformation & TwoDCurveDeformation: fixed the error message so it shows how many points will be in the snake.
snakeprogram3d.SnakeApplication updated version info.